modified:   backend/middleware/requireAuth.js 	modified:   frontend/src/components/WorkoutDetails.js 	modified:   frontend/src/components/WorkoutsForm.js 	modified:   frontend/src/pages/Home.js

+++ frontend/src/components/WorkoutDetails.js
@@ -1,14 +1,21 @@
 import React from "react";
 import { useWorkoutContext } from "../hooks/useWorkoutsContext";
-
+import { useAuthContext } from "../hooks/useAuthContext";
 // date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
+  const { user } = useAuthContext();
   const { dispatch } = useWorkoutContext();
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      }
     });
     const json = await response.json();
